+++ models/skillgroup.js
@@ -1,11 +1,11 @@
 module.exports = function (sequelize, DataTypes) {
     return sequelize.define("SkillGroup", {
         id: {
-            type: DataTypes.INTEGER,
+            type: DataTypes.BIGINT,
             primaryKey: true,
             autoIncrement: true
         },
-		group_id: { type: DataTypes.INTEGER},
-		skill_id : { type: DataTypes.INTEGER}
+		group_id: { type: DataTypes.BIGINT},
+		skill_id : { type: DataTypes.BIGINT}
     });
 };
