+++ models/skillquestion.js
@@ -1,12 +1,12 @@
 module.exports = function (sequelize, DataTypes) {
     return sequelize.define("SkillQuestion",  {
         id: {
-            type: DataTypes.INTEGER,
+            type: DataTypes.BIGINT,
             primaryKey: true,
             autoIncrement: true
         },
-		question_id : { type: DataTypes.INTEGER},
-		skill_id : { type: DataTypes.INTEGER}
+		question_id : { type: DataTypes.BIGINT },
+		skill_id : { type: DataTypes.BIGINT }
 		
     });
 };
